tests/wpt/web-platform-tests/svg/painting/parsing/stroke-miterlimit-invalid.svg

<?xml version="1.0" encoding="UTF-8"?>
<svg xmlns="http://www.w3.org/2000/svg"
     xmlns:h="http://www.w3.org/1999/xhtml"
     width="800px" height="800px">
  <title>SVG Painting: parsing stroke-miterlimit with invalid values</title>
  <metadata>
    <h:link rel="help" href="https://svgwg.org/svg2-draft/painting.html#StrokeMiterlimitProperty"/>
    <h:meta name="assert" content="stroke-miterlimit supports only the grammar 'number'"/>
    <h:meta name="assert" content="A negative value for stroke-miterlimit must be treated as an illegal value."/>
  </metadata>
  <g id="target"></g>
  <h:script src="/resources/testharness.js"/>
  <h:script src="/resources/testharnessreport.js"/>
  <h:script src="/css/support/parsing-testcommon.js"/>
  <script><![CDATA[

test_invalid_value("stroke-miterlimit", "1.");
test_invalid_value("stroke-miterlimit", "2 3");
test_invalid_value("stroke-miterlimit", "-4");

  ]]></script>
</svg>
